Created TypeBar for the BottomBar in chatBox
