+++ src/components/chat/w2wChat/TypeBar/Files/Files.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import './Files.css'
-import { FILE_ICON } from '../stickers/stickerHelper'
+import { FILE_ICON } from '../../stickers/stickerHelper'
 import { MessageIPFS } from 'helpers/w2w/ipfs'
 import { formatFileSize } from 'helpers/w2w'
