+++ src/components/chat/w2wChat/TypeBar/Gifs/GifPicker.tsx
@@ -6,7 +6,7 @@ import React, { FC, useRef, useState } from 'react';
 // Internal Compoonents
 import LoaderSpinner, { LOADER_TYPE } from 'components/reusables/loaders/LoaderSpinner';
 import { ItemVV2 } from "components/reusables/SharedStylingV2";
-import { useFetch } from '../stickers/useFetchHook';
+import { useFetch } from '../../stickers/useFetchHook';
 import { useClickAway } from 'hooks/useClickAway';
 import './GifPicker.css';
